JaegerExporter cleanup.
