+++ src/OpenTelemetry.Exporter.Jaeger/Implementation/EmitBatchArgs.cs
@@ -13,9 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 // </copyright>
-using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Thrift.Protocol;
@@ -23,56 +21,12 @@ using Thrift.Protocol.Entities;
 
 namespace OpenTelemetry.Exporter.Jaeger.Implementation
 {
-    internal class EmitBatchArgs : TUnionBase
+    internal class EmitBatchArgs
     {
-        public Batch Batch { get; set; }
-
-        public async Task WriteAsync(TProtocol oprot, CancellationToken cancellationToken)
+        public static async Task WriteAsync(int seqId, byte[] processMessage, List<BufferWriterMemory> spanMessages, TProtocol oprot, CancellationToken cancellationToken)
         {
-            oprot.IncrementRecursionDepth();
-            try
-            {
-                var struc = new TStruct("emitBatch_args");
-                await oprot.WriteStructBeginAsync(struc, cancellationToken);
-                if (this.Batch is Batch batch)
-                {
-                    var field = new TField
-                    {
-                        Name = "batch",
-                        Type = TType.Struct,
-                        ID = 1,
-                    };
+            await oprot.WriteMessageBeginAsync(new TMessage("emitBatch", TMessageType.Oneway, seqId), cancellationToken);
 
-                    await oprot.WriteFieldBeginAsync(field, cancellationToken);
-                    await batch.WriteAsync(oprot, cancellationToken);
-                    await oprot.WriteFieldEndAsync(cancellationToken);
-                }
-
-                await oprot.WriteFieldStopAsync(cancellationToken);
-                await oprot.WriteStructEndAsync(cancellationToken);
-            }
-            finally
-            {
-                oprot.DecrementRecursionDepth();
-            }
-        }
-
-        public override string ToString()
-        {
-            var sb = new StringBuilder("emitBatch_args(");
-
-            if (this.Batch is Batch batch)
-            {
-                sb.Append("Batch: ");
-                sb.Append(batch?.ToString() ?? "<null>");
-            }
-
-            sb.Append(")");
-            return sb.ToString();
-        }
-
-        internal static async Task WriteAsync(byte[] processMessage, List<BufferWriterMemory> spanMessages, TProtocol oprot, CancellationToken cancellationToken)
-        {
             oprot.IncrementRecursionDepth();
             try
             {
@@ -97,6 +51,9 @@ namespace OpenTelemetry.Exporter.Jaeger.Implementation
             {
                 oprot.DecrementRecursionDepth();
             }
+
+            await oprot.WriteMessageEndAsync(cancellationToken);
+            await oprot.Transport.FlushAsync(cancellationToken);
         }
     }
 }
